Delete legacy appengine code (#231)
- Delete all files behind `+build appengine`
- Merge the `*_vm.go` variants with the base files
- `go fix` for context update (TODO: Fix internal/api.go's use of context)
- Update travis config to test new versions
