+++ .travis.yml
@@ -10,9 +10,11 @@ script:
 
 matrix:
   include:
-    - go: 1.9.x
-      env: GOAPP=true
-    - go: 1.10.x
-      env: GOAPP=false
     - go: 1.11.x
       env: GO111MODULE=on
+    - go: 1.12.x
+      env: GO111MODULE=on
+    - go: 1.13.x
+    - go: 1.14.x
+    - go: 1.15.x
+    - go: 1.16.x
