+++ datastore/transaction.go
@@ -5,9 +5,8 @@
 package datastore
 
 import (
+	"context"
 	"errors"
-
-	"golang.org/x/net/context"
 
 	"google.golang.org/appengine/internal"
 	pb "google.golang.org/appengine/internal/datastore"
