+++ demos/guestbook/guestbook.go
@@ -2,17 +2,13 @@
 // Use of this source code is governed by the Apache 2.0
 // license that can be found in the LICENSE file.
 
-// This example only works on App Engine "flexible environment".
-// +build !appengine
-
 package main
 
 import (
+	"context"
 	"html/template"
 	"net/http"
 	"time"
-
-	"golang.org/x/net/context"
 
 	"google.golang.org/appengine"
 	"google.golang.org/appengine/datastore"
