+++ internal/main_test.go
@@ -1,5 +1,3 @@
-// +build !appengine
-
 package internal
 
 import (
