+++ log/api.go
@@ -7,7 +7,7 @@ package log
 // This file implements the logging API.
 
 import (
-	"golang.org/x/net/context"
+	"context"
 
 	"google.golang.org/appengine/internal"
 )
