+++ log/log.go
@@ -29,13 +29,13 @@ Example:
 package log // import "google.golang.org/appengine/log"
 
 import (
+	"context"
 	"errors"
 	"fmt"
 	"strings"
 	"time"
 
 	"github.com/golang/protobuf/proto"
-	"golang.org/x/net/context"
 
 	"google.golang.org/appengine"
 	"google.golang.org/appengine/internal"
