+++ namespace_test.go
@@ -5,9 +5,8 @@
 package appengine
 
 import (
+	"context"
 	"testing"
-
-	"golang.org/x/net/context"
 )
 
 func TestNamespaceValidity(t *testing.T) {
