+++ taskqueue/taskqueue.go
@@ -17,6 +17,7 @@ taskqueue operation is to add a single POST task, NewPOSTTask makes it easy.
 package taskqueue // import "google.golang.org/appengine/taskqueue"
 
 import (
+	"context"
 	"errors"
 	"fmt"
 	"net/http"
@@ -25,7 +26,6 @@ import (
 	"time"
 
 	"github.com/golang/protobuf/proto"
-	"golang.org/x/net/context"
 
 	"google.golang.org/appengine"
 	"google.golang.org/appengine/internal"
