+++ user/oauth.go
@@ -5,7 +5,7 @@
 package user
 
 import (
-	"golang.org/x/net/context"
+	"context"
 
 	"google.golang.org/appengine/internal"
 	pb "google.golang.org/appengine/internal/user"
